terraform/functions/chat/requirements.txt
boto3==1.35.76
PyJWT==2.10.1
